insights/requirements.txt
google-cloud-bigquery==3.34.0
db-dtypes==1.4.3
anthropic==0.52.0
